# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "提供所有API接口的说明文档",
        "title": "知时后端API文档",
        "contact": {
            "name": "吕舒君",
            "email": "Lvshujun0918@163.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/ping": {
            "get": {
                "description": "检查Gin和数据库服务的状态，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "summary": "获取服务状态",
                "responses": {
                    "200": {
                        "description": "服务状态检查结果，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        },
        "/user/info/{u_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定用户的信息，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "u_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户信息响应，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录获取JWT token，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录信息",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录响应，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        },
        "/user/logup": {
            "post": {
                "description": "新用户注册，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "registerRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.UserLogupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册响应，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        },
        "/v1/msg/post": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "提交一条消息记录，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "发布消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "u_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "消息内容",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.InternalUsualMsgPostReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息发布响应，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        },
        "/v1/report/{date}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据日期生成报告，始终返回HTTP 200，错误通过响应体中的errcode判断",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "生成报告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "u_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期 (格式: yyyy-mm-dd)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "报告生成响应，errcode=0表示成功",
                        "schema": {
                            "$ref": "#/definitions/internal.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal.InternalUsualMsgPostReq": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                }
            }
        },
        "internal.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据"
                },
                "errcode": {
                    "description": "错误码，0表示成功",
                    "type": "integer"
                },
                "message": {
                    "description": "消息内容",
                    "type": "string"
                }
            }
        },
        "route.UserLoginReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "route.UserLogupReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}